refactor(parsers): Split the output parsers by shape
`parsers.rs` becomes a directory module with one submodule per output *shape* rather than per tool. SARIF is the case that decides it: checkstyle already shares the format with cppcheck, SwiftLint, tflint and hadolint, so a further SARIF producer should cost a table entry and not a parser. The file was 532 lines against a 600-line soft limit with more parsers due; the largest module is now 201.

No behaviour change - `parse_output` dispatches to the same functions on the same format strings.

Also pins `json_kind_name`, which the move surfaced as untested: `json_parser_rejects_object_payload` asserted only that the error named the kind it *wanted*, so the half of the message saying what actually arrived passed under any answer. Two mutants survived on it. The new test walks all five JSON kinds.

Testing:
- 1 test added, 1,171 passing
- Scoped mutation sweep clean over the moved code
